Realign with EventStore-5.0.0-rc1

+++ src/EventStore.ClientAPI.Abstractions/StreamCheckpoint.cs
@@ -6,6 +6,6 @@
   public static class StreamCheckpoint
   {
     /// <summary>Indicates that a catch-up subscription should receive all events in the stream.</summary>
-    public static int? StreamStart = null;
+    public static long? StreamStart = null;
   }
 }
